requirements.txt
environs==11.0.0
python-telegram-bot==13.*
vk_api==11.9.9
redis==5.2.1